Fix conflict between dev-master and pull-request branches
Some of the builds fail because the circular dependencies "wp-cli -> external-command -> wp-cli" produce an unresolvable conflict.
The main issue is that we had required `"wp-cli/wp-cli": "dev-master"` in the external command packages so that they can always be tested against the latest version.
However, `dev-master` only means "the latest commit on the `master` branch". So, this would then exclude other `wp-cli/wp-cli` branches to be accepted ... like for example pull requests!
Current reasoning is that we just want to make sure that we have "any" WP-CLI to work with, and define the actual version to use for the tests through Travis. So, we think that adding the following combination to each external packages `composer.json` should do the trick:
```"minimum-stability": "dev",
"prefer-stable": true,
"require": {
   "wp-cli/wp-cli": "*"
}
```
This specific combination means: "use the *highest* available *stable* version of `wp-cli/wp-cli` while allowing to fall back to `dev-master` if the stable versions don't meet the requirements".